Dependency inversion: Create services layer
