+++ hq_superset/utils.py
@@ -1,15 +1,12 @@
 import ast
-import logging
-import os
 from contextlib import contextmanager
 from datetime import date, datetime
 from zipfile import ZipFile
 
 import pandas
 import sqlalchemy
-from flask import current_app, g
+from flask import current_app
 from flask_login import current_user
-from sqlalchemy.dialects import postgresql
 from superset.utils.database import get_or_create_db
 
 from .const import HQ_DATA
@@ -18,9 +15,6 @@ from .models import DataSetChange
 DOMAIN_PREFIX = "hqdomain_"
 SESSION_USER_DOMAINS_KEY = "user_hq_domains"
 SESSION_OAUTH_RESPONSE_KEY = "oauth_response"
-ASYNC_DATASOURCE_IMPORT_LIMIT_IN_BYTES = 5_000_000  # ~5MB
-
-logger = logging.getLogger(__name__)
 
 
 def get_datasource_export_url(domain, datasource_id):
@@ -103,39 +97,6 @@ def parse_date(date_str):
         return date_str
 
 
-class AsyncImportHelper:
-    def __init__(self, domain, datasource_id):
-        self.domain = domain
-        self.datasource_id = datasource_id
-
-    @property
-    def progress_key(self):
-        return f"{self.domain}_{self.datasource_id}_import_task_id"
-
-    @property
-    def task_id(self):
-        from superset.extensions import cache_manager
-
-        return cache_manager.cache.get(self.progress_key)
-
-    def is_import_in_progress(self):
-        if not self.task_id:
-            return False
-        from celery.result import AsyncResult
-        res = AsyncResult(self.task_id)
-        return not res.ready()
-
-    def mark_as_in_progress(self, task_id):
-        from superset.extensions import cache_manager
-
-        cache_manager.cache.set(self.progress_key, task_id)
-
-    def mark_as_complete(self):
-        from superset.extensions import cache_manager
-
-        cache_manager.cache.delete(self.progress_key)
-
-
 class DomainSyncUtil:
 
     def __init__(self, security_manager):
@@ -192,66 +153,6 @@ def get_datasource_file(path):
         yield zipfile.open(filename)
 
 
-def download_datasource(domain, datasource_id):
-    import superset
-    from hq_superset.hq_requests import HQRequest, HqUrl
-
-    hq_request = HQRequest(
-        url=HqUrl.datasource_export_url(domain, datasource_id),
-    )
-    response = hq_request.get()
-    if response.status_code != 200:
-        raise CCHQApiException("Error downloading the UCR export from HQ")
-
-    filename = f"{datasource_id}_{datetime.now()}.zip"
-    path = os.path.join(superset.config.SHARED_DIR, filename)
-    with open(path, "wb") as f:
-        f.write(response.content)
-
-    return path, len(response.content)
-
-
-def subscribe_to_hq_datasource(domain, datasource_id):
-    from superset.config import BASE_URL
-    from hq_superset.hq_requests import HQRequest, HqUrl
-    from hq_superset.models import HQClient
-
-    if HQClient.get_by_domain(domain) is None:
-        hq_request = HQRequest(
-            url=HqUrl.subscribe_to_datasource_url(domain, datasource_id)
-        )
-
-        client_id, client_secret = HQClient.create_domain_client(domain)
-
-        response = hq_request.post({
-            'webhook_url': f'{BASE_URL}/hq_webhook/change/',
-            'token_url': f'{BASE_URL}/oauth/token',
-            'client_id': client_id,
-            'client_secret': client_secret,
-        })
-        if response.status_code == 201:
-            return
-        if response.status_code < 500:
-            logger.error(
-                f"Failed to subscribe to data source {datasource_id} due to the following issue: {response.data}"
-            )
-        if response.status_code >= 500:
-            logger.exception(
-                f"Failed to subscribe to data source {datasource_id} due to a remote server error"
-            )
-
-
-def get_datasource_defn(domain, datasource_id):
-    from hq_superset.hq_requests import HQRequest, HqUrl
-
-    hq_request = HQRequest(url=HqUrl.datasource_details_url(domain, datasource_id))
-    response = hq_request.get()
-
-    if response.status_code != 200:
-        raise CCHQApiException("Error downloading the UCR definition from HQ")
-    return response.json()
-
-
 def convert_to_array(string_array):
     """
     Converts the string representation of a list to a list.
@@ -284,109 +185,6 @@ def convert_to_array(string_array):
         return []
 
     return array_values
-
-
-def refresh_hq_datasource(
-    domain,
-    datasource_id,
-    display_name,
-    file_path,
-    datasource_defn,
-    user_id=None,
-):
-    """
-    Pulls the data from CommCare HQ and creates/replaces the
-    corresponding Superset dataset
-    """
-    # See `CsvToDatabaseView.form_post()` in
-    # https://github.com/apache/superset/blob/master/superset/views/database/views.py
-
-    import superset
-    from superset import db
-    from superset.connectors.sqla.models import SqlaTable
-    from superset.sql_parse import Table
-
-    database = get_hq_database()
-    schema = get_schema_name_for_domain(domain)
-    csv_table = Table(table=datasource_id, schema=schema)
-    column_dtypes, date_columns, array_columns = get_column_dtypes(
-        datasource_defn
-    )
-
-    converters = {
-        column_name: convert_to_array for column_name in array_columns
-    }
-    # TODO: can we assume all array values will be of type TEXT?
-    sqlconverters = {
-        column_name: postgresql.ARRAY(sqlalchemy.types.TEXT)
-        for column_name in array_columns
-    }
-
-    def to_sql(df, replace=False):
-        database.db_engine_spec.df_to_sql(
-            database,
-            csv_table,
-            df,
-            to_sql_kwargs={
-                "if_exists": "replace" if replace else "append",
-                "dtype": sqlconverters,
-            },
-        )
-
-    try:
-        with get_datasource_file(file_path) as csv_file:
-
-            _iter = pandas.read_csv(
-                chunksize=10000,
-                filepath_or_buffer=csv_file,
-                encoding="utf-8",
-                parse_dates=date_columns,
-                date_parser=parse_date,
-                keep_default_na=True,
-                dtype=column_dtypes,
-                converters=converters,
-                iterator=True,
-                low_memory=True,
-            )
-
-            to_sql(next(_iter), replace=True)
-
-            for df in _iter:
-                to_sql(df, replace=False)
-
-        explore_database = get_explore_database(database)
-        sqla_table = (
-            db.session.query(SqlaTable)
-            .filter_by(
-                table_name=datasource_id,
-                schema=csv_table.schema,
-                database_id=explore_database.id,
-            )
-            .one_or_none()
-        )
-        if sqla_table:
-            sqla_table.description = display_name
-            sqla_table.fetch_metadata()
-        if not sqla_table:
-            sqla_table = SqlaTable(table_name=datasource_id)
-            # Store display name from HQ into description since
-            #   sqla_table.table_name stores datasource_id
-            sqla_table.description = display_name
-            sqla_table.database = explore_database
-            sqla_table.database_id = database.id
-            if user_id:
-                user = superset.appbuilder.sm.get_user_by_id(user_id)
-            else:
-                user = g.user
-            sqla_table.owners = [user]
-            sqla_table.user_id = user.get_id()
-            sqla_table.schema = csv_table.schema
-            sqla_table.fetch_metadata()
-            db.session.add(sqla_table)
-        db.session.commit()
-    except Exception as ex:  # pylint: disable=broad-except
-        db.session.rollback()
-        raise ex
 
 
 def get_explore_database(database):
